requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-ckeditor-5==0.2.11
django-jazzmin==3.0.1
Pillow==11.2.1
psycopg2==2.9.10
psycopg==3.2.9
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
drf-yasg==1.21.10
psycopg2-binary==2.9.9 

    